requirements.txt
paho-mqtt
pyModbusTCP
yaml8
schedule
